+++ src/tool/python/settings.h
@@ -1,6 +1,6 @@
 #pragma once
 
-namespace xlang
+namespace pywinrt
 {
     struct settings_type
     {
@@ -12,7 +12,7 @@ namespace xlang
 
         std::set<std::string> include;
         std::set<std::string> exclude;
-        meta::reader::filter filter;
+        xlang::meta::reader::filter filter;
     };
 
     extern settings_type settings;
